Presentacion v1.0.0

+++ FrontEnd/Prueba/src/app/vistas/bien/listar-bien/listar-bien.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Bien } from 'src/app/models/bien';
-import { Cliente } from 'src/app/models/cliente';
 import { BienService } from 'src/app/service/bien.service';
-import { ClienteService } from 'src/app/service/cliente.service';
 import { Router } from '@angular/router';
 
 @Component({
@@ -13,30 +11,46 @@ import { Router } from '@angular/router';
 export class ListarBienComponent implements OnInit {
 
   bien!: Bien[];
-
-  cliente: Cliente[] = [];
+  producto!: Bien[];
+  servicio!: Bien[];
 
   constructor(
     private service: BienService,
-    private clienteService: ClienteService,
     private router:Router) { }
 
   ngOnInit(): void{
     this.service.getBien().subscribe(data=>{
       this.bien=data;
     })
+    this.getProducto();
+    this.getServicio();
   }
 
-  Editar(bien:Bien):void{
+  getProducto(){
+    this.service.getProducto().subscribe(data=>{
+      this.producto=data;
+    })
+  }
+
+  getServicio(){
+    this.service.getServicio().subscribe(data=>{
+      this.servicio=data;
+    })
+  }
+
+  Editar(bien:Bien){
     localStorage.setItem("id",bien.id!.toString());
     this.router.navigate(["editar-bien"]);
   }
 
   Delete(bien:Bien){
-    this.service.deleteBien(bien).subscribe(data=>{
+    if(window.confirm('¿Estas seguro de querer borrar?')){
+      this.service.deleteBien(bien).subscribe(data=>{
       this.bien=this.bien?.filter(b=>b!==bien);
       alert("Bien eliminado!!!")
     })
+    }
+
   }
   Volver(){
     this.router.navigate(["proyecto"])
